+++ packages/policy/src/cli.ts
@@ -20,12 +20,14 @@ import {Octokit} from '@octokit/rest';
 import {exportToBigQuery} from './export';
 import {getPolicy, GitHubRepo} from './policy';
 import {getChanger} from './changer';
+import {openIssue} from './issue';
 
 interface Flags {
   repo?: string;
   search?: string;
   export?: boolean;
   autofix?: boolean;
+  report?: boolean;
 }
 
 const cli = meow(
@@ -38,9 +40,10 @@ const cli = meow(
     --search      Provide a GitHub repository search filter to limit the repositories used
     --export      Export the results to BigQuery.
     --autofix     Where possible, submit a PR to fix any issues.
+    --report      Create or update a GitHub Issue documenting the gaps
 
 	Examples
-    $ policy [--repo][--search QUERY][--autofix]
+    $ policy [--repo][--search QUERY][--autofix][--report]
 
 `,
   {
@@ -48,22 +51,21 @@ const cli = meow(
       repo: {type: 'string'},
       search: {type: 'string'},
       autofix: {type: 'boolean'},
+      report: {type: 'boolean'},
     },
   }
 );
 
 export async function main(cli: meow.Result<{}>) {
-  const token = process.env.GITHUB_TOKEN || process.env.GH_TOKEN;
+  const auth = process.env.GITHUB_TOKEN || process.env.GH_TOKEN;
   const flags = cli.flags as Flags;
-  if (!token) {
+  if (!auth) {
     throw new Error(
       'The GITHUB_TOKEN or GH_TOKEN env var must be set with a personal access token'
     );
   }
   const repos: GitHubRepo[] = [];
-  const octokit = new Octokit({
-    auth: token,
-  });
+  const octokit = new Octokit({auth});
   const policy = getPolicy(octokit, console);
   if (flags.repo) {
     const repo = await policy.getRepo(flags.repo);
@@ -84,6 +86,9 @@ export async function main(cli: meow.Result<{}>) {
     if (flags.autofix) {
       const changer = getChanger(octokit, repo);
       await changer.submitFixes(res);
+    }
+    if (flags.report) {
+      await openIssue(octokit, res);
     }
   }
 }